+++ app/src/main/java/com/lifespandh/ireflexions/utils/network/aws/S3UploadWorker.kt
@@ -9,6 +9,7 @@ import androidx.work.WorkerParameters
 import com.amplifyframework.core.Amplify
 import com.lifespandh.ireflexions.utils.file.createNewFile
 import com.lifespandh.ireflexions.utils.image.deserializeFromJson
+import com.lifespandh.ireflexions.utils.logs.logE
 import com.lifespandh.ireflexions.utils.network.LiveSubject
 import com.lifespandh.ireflexions.utils.network.UploadFileStatus
 import java.io.File
@@ -41,7 +42,7 @@ class S3UploadWorker(private val context: Context, workerParameters: WorkerParam
         }, {
             LiveSubject.FILE_UPLOAD_FILE.onNext(UploadFileStatus.Error(it))
         }).setOnProgress {
-            LiveSubject.FILE_UPLOAD_FILE.onNext(UploadFileStatus.FileStatus(it.fractionCompleted.toInt()))
+            LiveSubject.FILE_UPLOAD_FILE.onNext(UploadFileStatus.FileStatus(it.fractionCompleted))
         }
     }
